+++ src/dftbp/dftbplus/parser.F90
@@ -82,9 +82,10 @@ module dftbp_dftbplus_parser
   use dftbp_io_ipisocket, only : IPI_PROTOCOLS
 #:endif
 #:if WITH_TRANSPORT
-  use dftbp_transport_negfvars, only : TTransPar, TNEGFGreenDensInfo, TNEGFTunDos, TElPh,&
+  use dftbp_transport_negfvars, only : TNEGFGreenDensInfo, TNEGFTunDos, TElPh,&
       & ContactInfo
 #:endif
+  use dftbp_transport_negfvars, only : TTransPar
   implicit none
 
   private